+++ classes/suggested-tasks/providers/class-blog-description.php
@@ -10,7 +10,7 @@ namespace Progress_Planner\Suggested_Tasks\Providers;
 /**
  * Add tasks for Core blogdescription.
  */
-class Blog_Description extends Tasks {
+class Blog_Description extends Tasks_Interactive {
 
 	/**
 	 * Whether the task is an onboarding task.
@@ -25,6 +25,13 @@ class Blog_Description extends Tasks {
 	 * @var string
 	 */
 	protected const PROVIDER_ID = 'core-blogdescription';
+
+	/**
+	 * The popover ID.
+	 *
+	 * @var string
+	 */
+	const POPOVER_ID = 'core-blogdescription';
 
 	/**
 	 * Get the task title.
@@ -76,5 +83,40 @@ class Blog_Description extends Tasks {
 	 */
 	public function should_add_task() {
 		return '' === \get_bloginfo( 'description' );
+	}
+
+	/**
+	 * Get the popover instructions.
+	 *
+	 * @return void
+	 */
+	public function print_popover_instructions() {
+		?>
+		<p><?php \esc_html_e( 'In a few words, explain what this site is about. This information is used in your website\'s schema and RSS feeds, and can be displayed on your site. The tagline typically is your site\'s mission statement.', 'progress-planner' ); ?></p>
+		<?php
+	}
+
+	/**
+	 * Print the popover input field for the form.
+	 *
+	 * @return void
+	 */
+	public function print_popover_form_contents() {
+		?>
+		<label>
+			<p><?php echo \wp_kses_post( $this->get_task_details()['description'] ); ?></p>
+			<input
+				name="blogdescription"
+				type="text"
+				id="blogdescription"
+				value="<?php echo \esc_attr( \get_bloginfo( 'description' ) ); ?>"
+				class="regular-text"
+				placeholder="<?php \esc_html_e( 'A catchy phrase to describe your website', 'progress-planner' ); ?>"
+			>
+		</label>
+		<button type="submit" class="prpl-button prpl-button-primary" style="color: #fff;" disabled>
+			<?php \esc_html_e( 'Save', 'progress-planner' ); ?>
+		</button>
+		<?php
 	}
 }