+++ classes/suggested-tasks/providers/class-update-term-description.php
@@ -190,23 +190,6 @@ class Update_Term_Description extends Tasks {
 	}
 
 	/**
-	 * Transform data collector data into task data format.
-	 *
-	 * @param array $data The data from data collector.
-	 * @return array The transformed data with original data merged.
-	 */
-	protected function transform_collector_data( array $data ): array {
-		return \array_merge(
-			$data,
-			[
-				'target_term_id'   => $data['term_id'],
-				'target_taxonomy'  => $data['taxonomy'],
-				'target_term_name' => $data['name'],
-			]
-		);
-	}
-
-	/**
 	 * Get an array of tasks to inject.
 	 *
 	 * @return array
@@ -216,11 +199,11 @@ class Update_Term_Description extends Tasks {
 			return [];
 		}
 
-		$data    = $this->get_data_collector()->collect();
+		$data    = $this->transform_collector_data( $this->get_data_collector()->collect() );
 		$task_id = $this->get_task_id(
 			[
-				'term_id'  => $data['term_id'],
-				'taxonomy' => $data['taxonomy'],
+				'target_term_id'  => $data['target_term_id'],
+				'target_taxonomy' => $data['target_taxonomy'],
 			]
 		);
 
@@ -231,15 +214,12 @@ class Update_Term_Description extends Tasks {
 		// Transform the data to match the task data structure.
 		$task_data = $this->modify_injection_task_data(
 			$this->get_task_details(
-				$this->transform_collector_data( $data )
+				$data
 			)
 		);
 
-		// Get the task post.
-		$task_post = \progress_planner()->get_suggested_tasks_db()->get_post( $task_data['task_id'] );
-
 		// Skip the task if it was already injected.
-		if ( $task_post ) {
+		if ( \progress_planner()->get_suggested_tasks_db()->get_post( $task_data['task_id'] ) ) {
 			return [];
 		}
 
@@ -299,13 +279,10 @@ class Update_Term_Description extends Tasks {
 		}
 
 		$this->completed_term_ids = [];
-		$tasks                    = \progress_planner()->get_suggested_tasks_db()->get_tasks_by( [ 'provider_id' => $this->get_provider_id() ] );
 
-		if ( ! empty( $tasks ) ) {
-			foreach ( $tasks as $task ) {
-				if ( 'trash' === $task->post_status ) {
-					$this->completed_term_ids[] = $task->target_term_id;
-				}
+		foreach ( \progress_planner()->get_suggested_tasks_db()->get_tasks_by( [ 'provider_id' => $this->get_provider_id() ] ) as $task ) {
+			if ( 'trash' === $task->post_status ) {
+				$this->completed_term_ids[] = $task->target_term_id;
 			}
 		}