+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+= 1.7.0 =
+
+Bugs we fixed:
+
+* Fix issue where "Perform all updates" task was incorrectly marked as completed.
+* Prevent new lines to be added when user task title is edited.
+
+
 = 1.6.3 =
 
 Bugs we fixed: